Rearchitect image and context types
