+++ Sources/Graphics/context/GraphicsContextError.swift
@@ -1,0 +1,4 @@
+public enum GraphicsContextError: Error {
+    case couldNotCreate(width: Int, height: Int)
+    case couldNotMakeImage
+}
