+++ Sources/Graphics/image/ImageError.swift
@@ -1,0 +1,3 @@
+public enum ImageError: Error {
+    case couldNotCreate(width: Int, height: Int)
+}
